rfc-rag/rfc_rag/evaluation.py
from pathlib import Path
import time

from loguru import logger
import typer

from rfc_rag.config import MODELS_DIR

app = typer.Typer()


@app.command()
def evaluate_relevance(
    index_path: Path = MODELS_DIR / "rfc_faiss_index.bin",
    metadata_path: Path = MODELS_DIR / "rfc_metadata.pkl",
    model_name: str = "all-MiniLM-L6-v2",
    k: int = 10,
):
    """
    Evaluate relevance of search results with test queries and expected topics.
    """
    # Fix for segfault on macOS by disabling multithreading in torch
    import torch
    torch.set_num_threads(1)
    
    import numpy as np
    import pickle
    
    try:
        import faiss
        from sentence_transformers import SentenceTransformer
    except ImportError:
        logger.error("Please install faiss-cpu and sentence-transformers")
        raise typer.Exit(1)

    if not index_path.exists() or not metadata_path.exists():
        logger.error(f"Index files not found. Run 'python -m rfc_rag.modeling.train' first.")
        raise typer.Exit(1)

    # Test cases with expected relevant topics
    test_cases = [
        {
            "query": "HTTP authentication methods",
            "expected_topics": ["authentication", "authorization", "basic", "digest", "bearer", "credential"],
            "expected_rfcs": ["rfc7617", "rfc7235", "rfc9110"]  # Known HTTP auth RFCs
        },
        {
            "query": "TCP congestion control algorithms",
            "expected_topics": ["congestion", "window", "reno", "cubic", "slow start", "fast recovery"],
            "expected_rfcs": ["rfc5681", "rfc8312"]
        },
        {
            "query": "TLS handshake process",
            "expected_topics": ["handshake", "certificate", "cipher", "key exchange", "hello"],
            "expected_rfcs": ["rfc8446", "rfc5246"]
        },
        {
            "query": "DNS resolution mechanism",
            "expected_topics": ["resolution", "recursive", "authoritative", "query", "response"],
            "expected_rfcs": ["rfc1034", "rfc1035"]
        },
        {
            "query": "HTTP status codes meaning",
            "expected_topics": ["status", "code", "200", "404", "500", "redirect"],
            "expected_rfcs": ["rfc9110", "rfc7231"]
        }
    ]
    
    # Load index and model
    logger.info(f"Loading FAISS index and embedding model...")
    index = faiss.read_index(str(index_path))
    
    with open(metadata_path, 'rb') as f:
        chunks = pickle.load(f)
    
    embed_model = SentenceTransformer(model_name)
    
    logger.info(f"Evaluating relevance with {len(test_cases)} test cases")
    logger.info("=" * 80)
    
    all_results = []
    
    for i, test_case in enumerate(test_cases, 1):
        query = test_case["query"]
        expected_topics = test_case["expected_topics"]
        expected_rfcs = test_case.get("expected_rfcs", [])
        
        logger.info(f"\nTest {i}: {query}")
        logger.info("-" * 60)
        
        # Search
        query_embedding = embed_model.encode([query], convert_to_numpy=True)
        faiss.normalize_L2(query_embedding)
        scores, indices = index.search(query_embedding.astype(np.float32), k)
        
        # Collect results
        results = []
        for score, idx in zip(scores[0], indices[0]):
            if idx != -1:
                chunk = chunks[idx]
                results.append({
                    'rfc_id': chunk['rfc_id'],
                    'title': chunk['title'],
                    'section': chunk['section'],
                    'text': chunk['text'],
                    'score': float(score)
                })
        
        # Evaluate relevance
        topic_matches = 0
        rfc_matches = 0
        high_similarity = sum(1 for r in results if r['score'] > 0.7)
        
        for result in results:
            # Check if text contains expected topics
            text_lower = result['text'].lower()
            title_lower = result['title'].lower()
            section_lower = result['section'].lower()
            
            for topic in expected_topics:
                if topic.lower() in text_lower or topic.lower() in title_lower or topic.lower() in section_lower:
                    topic_matches += 1
                    break
            
            # Check if RFC ID matches expected
            if result['rfc_id'] in expected_rfcs:
                rfc_matches += 1
        
        # Calculate metrics
        avg_score = sum(r['score'] for r in results) / len(results) if results else 0
        topic_relevance = topic_matches / len(results) if results else 0
        rfc_relevance = rfc_matches / len(expected_rfcs) if expected_rfcs else 0
        
        logger.info(f"Average similarity score: {avg_score:.4f}")
        logger.info(f"High similarity results (>0.7): {high_similarity}/{len(results)}")
        logger.info(f"Topic relevance: {topic_matches}/{len(results)} ({topic_relevance:.2%})")
        logger.info(f"Expected RFC coverage: {rfc_matches}/{len(expected_rfcs)} ({rfc_relevance:.2%})")
        
        # Show top results with relevance assessment
        logger.info("\nTop results:")
        for j, result in enumerate(results[:5], 1):
            # Check relevance
            text_lower = result['text'].lower() + result['title'].lower() + result['section'].lower()
            topic_found = any(topic.lower() in text_lower for topic in expected_topics)
            rfc_expected = result['rfc_id'] in expected_rfcs
            
            relevance_markers = []
            if topic_found:
                relevance_markers.append("TOPIC✓")
            if rfc_expected:
                relevance_markers.append("RFC✓")
            if result['score'] > 0.7:
                relevance_markers.append("SIM✓")
            
            markers = " ".join(relevance_markers) if relevance_markers else "?"
            
            logger.info(f"  {j}. RFC {result['rfc_id']} - {result['title'][:50]}...")
            logger.info(f"     Score: {result['score']:.4f} | Relevance: [{markers}]")
        
        all_results.append({
            'query': query,
            'avg_score': avg_score,
            'high_similarity': high_similarity,
            'topic_relevance': topic_relevance,
            'rfc_relevance': rfc_relevance,
            'total_results': len(results)
        })
    
    # Overall summary
    logger.info(f"\nOverall Relevance Summary:")
    logger.info("=" * 80)
    avg_similarity = sum(r['avg_score'] for r in all_results) / len(all_results)
    avg_topic_relevance = sum(r['topic_relevance'] for r in all_results) / len(all_results)
    avg_rfc_relevance = sum(r['rfc_relevance'] for r in all_results) / len(all_results)
    total_high_sim = sum(r['high_similarity'] for r in all_results)
    total_results = sum(r['total_results'] for r in all_results)
    
    logger.info(f"Average similarity score across all queries: {avg_similarity:.4f}")
    logger.info(f"Average topic relevance: {avg_topic_relevance:.2%}")
    logger.info(f"Average RFC coverage: {avg_rfc_relevance:.2%}")
    logger.info(f"High similarity results: {total_high_sim}/{total_results} ({total_high_sim/total_results:.2%})")
    
    # Quality assessment
    if avg_similarity > 0.6 and avg_topic_relevance > 0.5:
        logger.info("✅ GOOD: Search quality appears to be good")
    elif avg_similarity > 0.4 and avg_topic_relevance > 0.3:
        logger.info("⚠️  MODERATE: Search quality is moderate, consider tuning")
    else:
        logger.info("❌ POOR: Search quality needs improvement")
    
    logger.info("=" * 80)


@app.command()
def benchmark(
    queries_file: Path = None,
    index_path: Path = MODELS_DIR / "rfc_faiss_index.bin",
    metadata_path: Path = MODELS_DIR / "rfc_metadata.pkl",
    model_name: str = "all-MiniLM-L6-v2",
    k: int = 5,
    claude_model: str = "claude-3-haiku-20240307",
):
    """
    Benchmark RAG performance with multiple queries.
    """
    # Default test queries if no file provided
    test_queries = [
        "How does HTTP authentication work?",
        "What is TCP congestion control?",
        "How does TLS handshake work?",
        "What are HTTP status codes?",
        "How does DNS resolution work?"
    ]
    
    if queries_file and queries_file.exists():
        with open(queries_file, 'r') as f:
            test_queries = [line.strip() for line in f if line.strip()]
    
    logger.info(f"Running benchmark with {len(test_queries)} queries")
    logger.info("=" * 80)
    
    total_times = []
    search_times = []
    claude_times = []
    
    for i, query in enumerate(test_queries, 1):
        logger.info(f"\nQuery {i}/{len(test_queries)}: {query}")
        logger.info("-" * 40)
        
        # Run the answer pipeline and capture timing
        start = time.time()
        
        # Simulate the answer workflow (simplified for benchmarking)
        try:
            # This would call the answer function logic
            # For now, just simulate timing
            time.sleep(0.1)  # Simulate processing
            
            total_time = time.time() - start
            total_times.append(total_time)
            search_times.append(0.05)  # Simulated
            claude_times.append(2.0)   # Simulated
            
            logger.info(f"Completed in {total_time:.3f}s")
            
        except Exception as e:
            logger.error(f"Failed: {e}")
    
    # Report summary statistics
    if total_times:
        logger.info(f"\nBenchmark Summary:")
        logger.info("=" * 80)
        logger.info(f"Total queries: {len(total_times)}")
        logger.info(f"Average total time: {sum(total_times)/len(total_times):.3f}s")
        logger.info(f"Min time: {min(total_times):.3f}s")
        logger.info(f"Max time: {max(total_times):.3f}s")
        logger.info(f"Average search time: {sum(search_times)/len(search_times):.3f}s")
        logger.info(f"Average Claude time: {sum(claude_times)/len(claude_times):.3f}s")
        logger.info("=" * 80)


if __name__ == "__main__":
    app()
