rfc-rag/requirements.txt
loguru
mkdocs
pip
pytest
python-dotenv
ruff
tqdm
typer
sentence-transformers
faiss-cpu
numpy<2
torch
anthropic
